Add to PerspectiveController
Uses a enum to save the current state -> don't have to pass the name
anymore.

To change state, passes a name, the frame and MenuController. The name
is checked and other parameters passed to the correct private method
which enables wanted MenuItems, adds the view to the frame, and sets the
current state to the selected one.

To update the state, passes the information. The current state is
checked and information passed to the correct private method which does
whatever it needs to with the information.

Saves the menubar so it can be put back onto the frame (removing all
components easier than finding all the elements added by the
perspective).

Change the controller to using the PerspectiveController rather than the
ViewController.

Edit the controller so changing a panel involves removing all
components, putting the menubar back on, than changing the perspective's
state.