Change PerspectiveController
Requires a MenuController when Initialised - is passed to perspectives
to add menuitems and enable menuitems.
Is passed Perspectives and names through a method - saved to a map where
the name maps to the Perspective.
A JPanel is passed back to the MainController to add to the JFrame. The
removes need to add MenuBar back on after each Perspective.
Perspectives are changed by name - if name doesn't exist or points to
null error is thrown. Perspective passes a JPanel through the enable
method which is put into the JPanel and can be easily removed.
Update class allows Perspectives to change Perspectives - ie. from
prelaunch to launch can only happen if certain events happen in
prelaunch.